move VMMemory to a separate file
